+++ src/javascripts/maps/maps.js
@@ -1,20 +1,19 @@
-import { add_grid, get_grid, get_table_range } from "../common/grid.js";
-import { set_palette, get_palette } from "../common/palette.js";
+import { add_grid, get_grid} from "../common/grid.js";
+import { set_palette} from "../common/palette.js";
 import {
   mapSettings,
   dimensions,
   getMapDataAttributes,
   legendSettings,
 } from "./config.js";
-import { get_data_level } from "../common/utils.js";
-import { d3, colours, msgBox } from "/bundle.js";
-import { data_check } from "/bundle.js";
-import { createTable } from "../common/errorbox.js";
+import { d3, colours} from "/bundle.js";
 import { add_spinner } from "../common/spinner.js";
-const check_duplicate_rows = data_check.check_duplicate_rows;
-const DuplicateRow = data_check.DuplicateRow;
-const EcodeParseError = data_check.EcodeParseError;
 import { graph_tools } from "/bundle.js";
+import { map_choose_scale_function,
+         match_geojson_to_datalookup,
+         validate_rows,
+         validate_data_level
+     } from "./common_maps.js";
 const VerticalLegend = graph_tools.VerticalLegend;
 const download_svg = graph_tools.download_svg;
 
@@ -101,35 +100,10 @@ function downloadAndProcessMapData(
   // Make loader to wait for map download
   add_spinner(container);
   d3.json(file).then((map) => {
-    // Keep track of rows matched
-    const usedKeys = new Set();
-    const dataFeatures = map.features.flatMap((feature) => {
-      const matchedData = dataLookup[feature.properties[ecodeid]];
-      if (matchedData) {
-        usedKeys.add(feature.properties[ecodeid]);
-        return { ...feature, data: { ...matchedData } };
-      } else {
-        return feature;
-      }
-    });
-
-    const unmatchedEcodes = Object.keys(dataLookup).filter(
-      (key) => !usedKeys.has(key)
-    );
-    if (unmatchedEcodes.length > 0) {
-      const lst = document.createElement("ul");
-      for (let r in unmatchedEcodes) {
-        const el = document.createElement("li");
-        el.textContent = unmatchedEcodes[r];
-        lst.appendChild(el);
-      }
-      const errorbox = document.createElement("div");
-      const text = document.createElement("p");
-      text.textContent = "You have entered unrecognised Ecode(s)";
-      errorbox.appendChild(text);
-      errorbox.appendChild(lst);
-      msgBox("Unrecognised Ecode(s)", errorbox);
-      container.innerHTML = "";
+    const dataFeatures = match_geojson_to_datalookup(map, dataLookup, ecodeid)
+    if (dataFeatures == -1) {
+        container.innerHTML = "";
+        return -1
     }
     // Make tooltip
     const tooltip = document.createElement("div");
@@ -325,125 +299,19 @@ export function draw_map(
   const dataLookup = createDataLookup(data);
   container.innerHTML = "";
   // Validate data
-  try {
-    check_duplicate_rows(data, "ecode");
-  } catch (error) {
-    if (error instanceof DuplicateRow) {
-      const lst = document.createElement("ul");
-      for (let r in error.rows) {
-        const el = document.createElement("li");
-        el.textContent = error.rows[r];
-        lst.appendChild(el);
-      }
-      const errorbox = document.createElement("div");
-      const text = document.createElement("p");
-      text.textContent = "You have entered duplicate Ecodes";
-      errorbox.appendChild(text);
-      errorbox.appendChild(lst);
-      msgBox("Duplicate Ecode", errorbox);
-      container.innerHTML = "";
-      return -1;
-    }
+  if (validate_rows(data) == -1) {
+    return -1
   }
-
-  let data_level;
-  try {
-    data_level = get_data_level(dataLookup, la_level);
-  } catch (error) {
-    if (error instanceof EcodeParseError) {
-      const tbl = createTable(error.ecodes);
-      const errorbox = document.createElement("div");
-      const text = document.createElement("p");
-      text.textContent =
-        "You have entered Ecodes from both counties and districts";
-      errorbox.appendChild(text);
-      errorbox.appendChild(tbl);
-      msgBox("Ecode Error", errorbox);
-      container.innerHTML = "";
-    }
-    return 0;
+  const data_level = validate_data_level(dataLookup, la_level);
+  if (data_level == -1) {
+    return -1
   }
   const { file, nameid, ecodeid } = getMapDataAttributes(data_level, data_year);
 
-  // create the scale
-  let scale;
-  let uniques;
-  // In the future this should be moved to get_map_page_state
-  const org_list = document.getElementById("org_list");
-  const palette = get_palette(org_list);
-  // For now everyone needs to use the light colours
-  const palette_size = Object.keys(palette.light).length;
-  const palette_elements = [...document.querySelectorAll(".palette-cell")];
-  // This is very inefficient
-  palette_elements.forEach((cell) => {
-    cell.classList.remove("palette-cell__disabled");
-  });
-  switch (data_check.guess_data_type(data)) {
-    case "float":
-      scale = d3.scaleSequential(
-        get_table_range(data),
-        d3.interpolate("rgb(0,0,0)", `rgb(${colour.join(",")})`)
-      );
-      break;
-    case "integer":
-      scale = d3.scaleSequential(
-        get_table_range(data),
-        d3.interpolate("rgb(0,0,0)", `rgb(${colour.join(",")})`)
-      );
-      break;
-    case "string":
-      uniques = data_check.get_data_uniques(data);
-      if (uniques.size <= palette_size) {
-        // This is definitely discrete. Should:
-        // 1 - Check if there are < 10 levels
-        // 2 - If so ask if they're ordered
-        // 3 - Make a scale based on the palette
-        // palette scale
-        // Get ordinal setting
-        const ordinalSetting = document.querySelector(
-          'input[name="ordinal-state"]:checked'
-        ).value;
-        if (ordinalSetting == "unordered") {
-          scale = d3.scaleOrdinal(
-            [...uniques],
-            Object.values(palette.light).map(
-              (colour) => `rgb(${colour.join(",")})`
-            )
-          );
-          // soft disable the palette
-          palette_elements.forEach((cell) => {
-            cell.classList.add("palette-cell__disabled");
-          });
-        } else {
-          // ordered scale
-          scale = d3.scaleOrdinal(
-            [...uniques],
-            Array.from({ length: uniques.size }, (_, i) =>
-              d3.interpolate(
-                "rgb(0,0,0)",
-                `rgb(${colour.join(",")})`
-              )((i + 1) / (uniques.size + 1))
-            )
-          );
-        }
-      } else {
-        // Message the user and tell them that they have too many levels
-        const lst = document.createElement("ul");
-        const unique_array = [...uniques];
-        for (let r in unique_array) {
-          const el = document.createElement("li");
-          el.textContent = unique_array[r];
-          lst.appendChild(el);
-        }
-        const errorbox = document.createElement("div");
-        const text = document.createElement("p");
-        text.textContent = `Cannot graph ordinal data with more than ${palette_size} levels. You have ${uniques.size}.`;
-        errorbox.appendChild(text);
-        errorbox.appendChild(lst);
-        msgBox("Too many ordinal levels", errorbox);
-        container.innerHTML = "";
-        return 0;
-      }
+  let scale = map_choose_scale_function(colour, data);
+  if (scale == -1) {
+    container.innerHTML = "";
+    return -1
   }
 
   downloadAndProcessMapData(
